FEAT: 랭킹 페이지 구조 개선 및 타입 분리
