+++ src/models/Ranking.model.ts
@@ -1,0 +1,8 @@
+export interface RankingData {
+  id: number;
+  username: string;
+  level: number;
+  like: number;
+  comments: number;
+  ranking: number;
+}
